move DI for msg sender away from program.cs
